requirements.txt
appdirs
asgiref
astroid
attrs
Babel
certifi
chardet
charset-normalizer
click
colorama
coverage
csscompressor
diff-cover
dill
Django
django-auth-ldap
django-cors-headers
django-pytest
django-simple-history
djangorestframework
flake8
ghp-import
htmlmin
htmlmin2
idna
iniconfig
isort
Jinja2
jsmin
lazy-object-proxy
Markdown
MarkupSafe
mccabe
mergedeep
mkdocs
mkdocs-material
mkdocs-material-extensions
mkdocs-minify-plugin
mssql-django
mypy
mypy-extensions
packaging
paginate
pathspec
platformdirs
pluggy
psycopg2
pyasn1
pyasn1-modules
pycodestyle
pyflakes
Pygments
pylint
pylint-django
pylint-plugin-utils
pymdown-extensions
pyodbc
pytest
pytest-cov
pytest-django
python-dateutil
python-ldap
pytz
PyYAML
pyyaml_env_tag
regex
requests
six
sqlfluff
sqlparse
tblib
toml
tomlkit
tqdm
typing_extensions
tzdata
urllib3
watchdog
wrapt
XlsxWriter
